+++ frontend/src/layout/Content/Content.tsx
@@ -1,5 +1,5 @@
-import { Grid, GridItemProps } from '../Grid';
 import { ContainerProps, LayoutProps, FlexContainerProps } from '@/types/style';
+import { Grid, GridItemProps } from '@/common/Grid';
 
 type ContentProps = {
   children?: React.ReactNode;
